+++ project/app/Http/Requests/CreateProjectRequest.php
@@ -14,10 +14,11 @@ class CreateProjectRequest extends FormRequest
     public function rules()
     {
         return [
-            'name'      => 'required|min:3',
-            'client_id' => 'required|exists:clients,id',
-            'lead'      => 'nullable|exists:resources,id',
-            'manager'   => 'nullable|exists:resources,id',
+            'name'       => 'required|min:3',
+            'innovation' => 'nullable|boolean',
+            'client_id'  => 'required|exists:clients,id',
+            'lead'       => 'nullable|exists:resources,id',
+            'manager'    => 'nullable|exists:resources,id',
         ];
     }
 }