+++ project/resources/views/dashboard.blade.php
@@ -91,31 +91,7 @@
 
     @livewire('clients.satisfaction', 'Clients satisfaits')
 
-    <div class="w-1/4 bg-white rounded shadow mx-4 px-4">
-      <div class="h-4"></div>
-
-      <div class="flex align-center">
-
-        <div class="w-20 flex items-center justify-center mr-2">
-          <span data-jam="battery-charging" data-width="48" data-height="48"
-                class="text-orange-500 fill-current"></span>
-        </div>
-
-        <div>
-          <p class="text-base text-gray-600">
-            Innovations
-          </p>
-          <p class="text-gray-800">
-          <span class="text-lg">
-            13
-          </span>
-          </p>
-        </div>
-
-      </div>
-
-      <div class="h-4"></div>
-    </div>
+    @livewire('projects.innovations', 'Innovations')
 
   </section>