+++ Changelog.md
@@ -24,10 +24,15 @@ We follow Semantic Versioning.
 # Table of contents
 
 - [Unreleased](#unreleased)
+- [v1.3.0](#v1.3.0)
+- [v1.2.0](#v1.2.0)
 - [v1.1.0](#v1.1.0)
 
-# Unreleased
+# v1.3.0
 
+- `Fixed` users can now access project and client from tasks tables
+- `Added` ability to mark project as 'innovation'
+- `Added` ability to filter projects by innovation
 - Changelog updated
 
 # v1.2.2